[ADM-887][frontend]: feat: move form initialization to MetricsStepper to persist the form submition state while jumping pages.
